Set up routing between search and compare
