+++ assets/src/js/components/compare/SearchParameterBreadcrumbs.vue
