+++ assets/src/js/components/search/SpaceSearch.vue
