+++ assets/src/js/components/search/TimeCardHeader.vue
